requirements.txt
absl-py==1.2.0
anyio==4.9.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
blinker==1.9.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
distlib==0.3.6
Django==4.1.1
exceptiongroup==1.0.4
filelock==3.8.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.49.1
h11==0.16.0
h5py==3.7.0
httpcore==1.0.9
httpx==0.26.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pillow==10.3.0
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
python-telegram-bot==20.8
pytube==15.0.0
pytube3==9.6.4
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.13.2
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
webdriver-manager==3.8.5
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
